requirements.txt
Django~=5.1
python-dotenv~=1.0.1
djangorestframework~=3.15.2
requests~=2.32.3
celery~=5.4.0
djangorestframework-simplejwt~=5.3.1